Code Cleanup and structuring.
