+++ src/routes/ml.routes.ts
@@ -1,8 +1,9 @@
 import express from "express";
 import { predictCategory } from "../controllers/ml.controller";
+import authenticate from "../middlewares/auth.middleware";
 
 const router = express.Router();
 
-router.post("/categorise-expense", predictCategory);
+router.post("/categorise-expense", authenticate, predictCategory);
 
 export default router;
