+++ src/types/ml.ts
@@ -1,0 +1,9 @@
+export type PredictRequest = {
+  title: string;
+  description: string;
+};
+
+export type PredictResponse = {
+  categoryId: string;
+  confidence: number;
+};
